pointer arithmetic

+++ src/main/resources-test/07 global vars.c
@@ -1,6 +1,6 @@
-u8 space = 0x20;
-u8 next = '?';
-u8* ptrToSpace = &space;
+i16 space = 0x20;
+i16 next = '?';
+i16* ptrToSpace = &space;
 
 void main() {
 	ptrToSpace = ptrToSpace + 1;
